Use 1.5 line spacing (Microsoft Word definition)
Spacing should be 1.5 times the single line spacing for examinations.
Note that this is larger than the one-half spacing given by the setspace
package due to the differing definitions. The definition used here
mirrors the one used by Microsoft Word for 1.5 spacing.

The spacing is set for the entire document (including abstract, contents
section and bibliography) but not for the title page. The title page
uses single spacing.