Use 10pt sans serif for captions and tables
Captions and tables should use 10pt and a sans serif font (Arial,
Helvetica or Calibri). Here we use Helvetica. For the font size, we make
use of 'footnotesize' which is 10pt for a 12pt document.

The font size applies to the entire table in accordance with the
Academic Program Policies.